+++ packages/bluesky/lib/bluesky.dart
@@ -260,6 +260,7 @@ export 'package:bluesky/src/entities/skeleton_feed.dart';
 export 'package:bluesky/src/entities/skeleton_feed_view.dart';
 export 'package:bluesky/src/entities/skeleton_reason.dart';
 export 'package:bluesky/src/entities/skeleton_reason_repost.dart';
+export 'package:bluesky/src/extension/grouped_notification_extension.dart';
 export 'package:bluesky/src/extension/at_uri_extension.dart';
 export 'package:bluesky/src/extension/blob_extension.dart';
 export 'package:bluesky/src/extension/strong_ref_extension.dart';